Rename: 파일 위치 수정
core로 파일 두개 옮김
