+++ src/main/java/com/ssafy/tott/api/core/APIResponse.java
@@ -1,4 +1,4 @@
-package com.ssafy.tott.api;
+package com.ssafy.tott.api.core;
 
 public interface APIResponse {
 }
