+++ src/main/java/com/ssafy/tott/api/kakao/data/dto/response/KakaoAPIResponse.java
@@ -1,7 +1,7 @@
 package com.ssafy.tott.api.kakao.data.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.ssafy.tott.api.APIResponse;
+import com.ssafy.tott.api.core.APIResponse;
 import com.ssafy.tott.api.kakao.data.vo.Documents;
 import lombok.Data;
