+++ src/main/java/com/ssafy/tott/api/seoulopendata/data/dto/response/RentAPIResponse.java
@@ -2,7 +2,7 @@ package com.ssafy.tott.api.seoulopendata.data.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.ssafy.tott.api.APIResponse;
+import com.ssafy.tott.api.core.APIResponse;
 import com.ssafy.tott.api.seoulopendata.data.vo.RentData;
 import lombok.Getter;
 import lombok.ToString;
