+++ src/main/java/com/ssafy/tott/api/shinhan/dto/response/ShinhanBankAPIResponse.java
@@ -1,7 +1,7 @@
 package com.ssafy.tott.api.shinhan.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.ssafy.tott.api.APIResponse;
+import com.ssafy.tott.api.core.APIResponse;
 import com.ssafy.tott.api.shinhan.dto.response.header.ResponseDataHeader;
 import lombok.AccessLevel;
 import lombok.Getter;
